Reduce complexity of panels endpoint by moving some code to the panels controllers (#2981)
* Reduce complexity of  endpoint moving some code to the panels controllers

* Add missing imports

* remove spaces

* Move other 2 functions to controllers

* A more precise comment
